# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BookPage": {
        "description": "One book page: parsed hadiths, or raw page text when none are parsed yet.",
        "properties": {
          "chapter_title": {
            "title": "Chapter Title",
            "type": "string"
          },
          "chapter_title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Title En"
          },
          "footnotes": {
            "description": "The page's footnote apparatus in printed order; empty when the source row carries no footnote block. Served for raw and hadith-parsed pages alike: the notes annotate the printed page, not the extraction.",
            "items": {
              "$ref": "#/components/schemas/Footnote"
            },
            "title": "Footnotes",
            "type": "array"
          },
          "hadiths": {
            "items": {
              "$ref": "#/components/schemas/Hadith"
            },
            "title": "Hadiths",
            "type": "array"
          },
          "page_number": {
            "minimum": 1.0,
            "title": "Page Number",
            "type": "integer"
          },
          "section_title": {
            "title": "Section Title",
            "type": "string"
          },
          "section_title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title En"
          },
          "text_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Raw page text in Arabic, present when the pipeline has not parsed the page into structured hadiths. Mutually exclusive with a populated hadiths list; never both.",
            "title": "Text Ar"
          },
          "text_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "English translation of the raw page text, paired with text_ar. None until the pipeline emits a translation for the page; the reader shows a labelled preview in that case and this real text the moment it arrives.",
            "title": "Text En"
          },
          "total_pages": {
            "minimum": 1.0,
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": ["page_number", "total_pages", "chapter_title", "section_title", "hadiths"],
        "title": "BookPage",
        "type": "object"
      },
      "BookSearchMatch": {
        "description": "One in-book search hit: the page and a short Arabic excerpt.",
        "properties": {
          "page": {
            "description": "Page the match occurs on.",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "snippet": {
            "description": "Excerpt of the matching unit (Arabic).",
            "title": "Snippet",
            "type": "string"
          }
        },
        "required": ["page", "snippet"],
        "title": "BookSearchMatch",
        "type": "object"
      },
...
      "CrossRef": {
        "description": "Pointer to a parallel narration in another collection.",
        "properties": {
          "book": {
            "description": "Source book (English transliteration).",
            "title": "Book",
            "type": "string"
          },
          "book_ar": {
            "description": "Source book (Arabic).",
            "title": "Book Ar",
            "type": "string"
          },
          "chapter": {
            "description": "Chapter or volume reference.",
            "title": "Chapter",
            "type": "string"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Page number in the source book.",
            "title": "Page"
          }
        },
        "required": ["book", "book_ar", "chapter"],
        "title": "CrossRef",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "description": "The body of every non-2xx response the API declares.",
        "properties": {
          "detail": {
            "description": "Human-readable failure description.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": ["detail"],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "Footnote": {
        "description": "One entry in a page's footnote apparatus: the editor's notes as printed.\n\nThe apparatus is page-local, mirroring the print edition: an entry is\nserved with the page whose foot it is printed on. Roughly 1% of corpus\nmarkers reference an entry printed on the neighboring page; those entries\nstill appear on their own page, exactly as the edition prints them.",
        "properties": {
          "marker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Entry number as printed in the edition ('1', '2', ...). None for unnumbered note text: a free-form editorial block, or the continuation of the previous page's entry in continuously numbered editions. The block split is lossless, so None never means a parse failure.",
            "title": "Marker"
          },
          "text": {
            "description": "Arabic note text as printed.",
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["text"],
        "title": "Footnote",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Hadith": {
        "description": "One hadith record: isnad + matn + narrators + cross-references.",
        "properties": {
          "cross_refs": {
            "description": "Parallel narrations.",
            "items": {
              "$ref": "#/components/schemas/CrossRef"
            },
            "title": "Cross Refs",
            "type": "array"
          },
          "grade": {
            "anyOf": [
              {
                "enum": ["sahih", "hasan", "daif", "mawdu"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Authenticity grade. None when the page is raw corpus text that the pipeline has not yet graded; the UI hides the grade pill in that case.",
            "title": "Grade"
          },
          "isnad_ar": {
            "description": "Chain of transmission in Arabic.",
            "title": "Isnad Ar",
            "type": "string"
          },
          "matn_ar": {
            "description": "Body of the hadith in Arabic.",
            "title": "Matn Ar",
            "type": "string"
          },
          "matn_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "English translation of the matn.",
            "title": "Matn En"
          },
          "n": {
            "description": "Sequence number on the page.",
            "minimum": 1.0,
            "title": "N",
            "type": "integer"
          },
          "narrators": {
            "description": "Narrators in the chain.",
            "items": {
              "$ref": "#/components/schemas/Narrator"
            },
            "title": "Narrators",
            "type": "array"
          }
        },
        "required": ["n", "isnad_ar", "matn_ar", "narrators"],
        "title": "Hadith",
        "type": "object"
      },
...
      "Narrator": {
        "description": "One narrator in an isnad.",
        "properties": {
          "d": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Death year, Hijri.",
            "title": "D"
          },
          "grade": {
            "description": "Biographical evaluation (Trustworthy, Reliable, ...).",
            "title": "Grade",
            "type": "string"
          },
          "name": {
            "description": "Name in romanized form.",
            "title": "Name",
            "type": "string"
          },
          "name_ar": {
            "description": "Name in Arabic.",
            "title": "Name Ar",
            "type": "string"
          },
          "person_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Registry link: id in /api/person, when the name resolved to an enriched person instead of a rijal entry. Mutually exclusive with rijal_id.",
            "title": "Person Id"
          },
          "rijal_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Registry link: id in /api/rijal, resolved at build time by normalized-name match. None when the registry does not know this narrator; the reader shows the name unlinked rather than guessing.",
            "title": "Rijal Id"
          },
          "role": {
            "description": "Role / position in the chain (companion, transmitter, ...).",
            "title": "Role",
            "type": "string"
          }
        },
        "required": ["name", "name_ar", "role", "grade"],
        "title": "Narrator",
        "type": "object"
      },
...
      "Page_BookSearchMatch_": {
        "properties": {
          "items": {
            "description": "Records in this slice.",
            "items": {
              "$ref": "#/components/schemas/BookSearchMatch"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Slice size requested.",
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Number of records skipped before this slice.",
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total records matching the query, across all pages.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "Page[BookSearchMatch]",
        "type": "object"
      },
...
      "Page_RijalEntry_": {
        "properties": {
          "items": {
            "description": "Records in this slice.",
            "items": {
              "$ref": "#/components/schemas/RijalEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Slice size requested.",
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Number of records skipped before this slice.",
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total records matching the query, across all pages.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "Page[RijalEntry]",
        "type": "object"
      },
...
      "RijalEntry": {
        "description": "One narrator in the rijal registry (reliability-graded).",
        "properties": {
          "birth_year": {
            "default": "",
            "description": "Birth year as recorded (Hijri, free-form).",
            "title": "Birth Year",
            "type": "string"
          },
          "book_path": {
            "default": "",
            "description": "Relative path of the source corpus file.",
            "title": "Book Path",
            "type": "string"
          },
          "category": {
            "default": "clean",
            "description": "Data-quality class (clean, editorial).",
            "title": "Category",
            "type": "string"
          },
          "death_year": {
            "default": "",
            "description": "Death year as recorded (Hijri, free-form).",
            "title": "Death Year",
            "type": "string"
          },
          "evaluator": {
            "default": "",
            "description": "Critic who issued the reliability term.",
            "title": "Evaluator",
            "type": "string"
          },
          "full_name": {
            "description": "Full name in Arabic.",
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "description": "Stable corpus index, also the detail-route key.",
            "minimum": 0.0,
            "title": "Id",
            "type": "integer"
          },
          "kunya": {
            "default": "",
            "description": "Teknonym (Abu/Umm ...), if recorded.",
            "title": "Kunya",
            "type": "string"
          },
          "nisba": {
            "default": "",
            "description": "Attributive name (tribe/place), if recorded.",
            "title": "Nisba",
            "type": "string"
          },
          "reliability_grade": {
            "default": "",
            "description": "Numeric reliability grade, as text.",
            "title": "Reliability Grade",
            "type": "string"
          },
          "reliability_term": {
            "default": "",
            "description": "Primary reliability term (thiqa, ...).",
            "title": "Reliability Term",
            "type": "string"
          },
          "source_label": {
            "default": "",
            "description": "Human label of the source work + volume.",
            "title": "Source Label",
            "type": "string"
          },
          "student_count": {
            "description": "Number of recorded students.",
            "minimum": 0.0,
            "title": "Student Count",
            "type": "integer"
          },
          "teacher_count": {
            "description": "Number of recorded teachers.",
            "minimum": 0.0,
            "title": "Teacher Count",
            "type": "integer"
          },
          "tradition": {
            "default": "",
            "description": "Sunni / shia / both, when classified.",
            "title": "Tradition",
            "type": "string"
          }
        },
        "required": ["full_name", "id", "teacher_count", "student_count"],
        "title": "RijalEntry",
        "type": "object"
      },
...
      "Toc": {
        "description": "The full table of contents for one book.",
        "properties": {
          "book_urn": {
            "description": "URN of the book this TOC belongs to.",
            "title": "Book Urn",
            "type": "string"
          },
          "entries": {
            "description": "Ordered TOC rows.",
            "items": {
              "$ref": "#/components/schemas/TocEntry"
            },
            "title": "Entries",
            "type": "array"
          }
        },
        "required": ["book_urn", "entries"],
        "title": "Toc",
        "type": "object"
      },
      "TocEntry": {
        "description": "One row in a book's table of contents.",
        "properties": {
          "active": {
            "default": false,
            "description": "Whether this is the section currently being read.",
            "title": "Active",
            "type": "boolean"
          },
          "page": {
            "description": "Page number where this section begins.",
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "title": {
            "description": "Arabic section title.",
            "title": "Title",
            "type": "string"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "English section title, if translated.",
            "title": "Title En"
          }
        },
        "required": ["page", "title"],
        "title": "TocEntry",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Knowledge-graph reader for classical Arabic manuscripts.",
    "title": "Shia Online Library",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/books/{book_urn}/pages/{page_number}": {
      "get": {
        "description": "Return the requested page or raise ``ResourceNotFoundError``.\n\nStructured ``hadiths`` (isnad + matn + narrators) come from the manuscript\nindex built by Phase 3 extract, via ``manuscript.hadiths_for_page``. When the\npage has parsed hadiths they are served and ``text_ar`` is ``None`` (the two\nare mutually exclusive per the ``BookPage`` contract); otherwise ``text_ar``\ncarries the raw page text honestly \u2014 for prose pages, or before the index is\nbuilt. ``text_en`` is the single wiring point for an English rendering: the\ncorpus has no English column yet, so it stays ``None`` and the reader shows a\nlabelled preview; set it here the moment translations land. ``footnotes``\ncarries the page's printed apparatus on both shapes: the notes annotate the\nprinted page, not the extraction.",
        "operationId": "get_page_api_books__book_urn__pages__page_number__get",
        "parameters": [
          {
            "in": "path",
            "name": "book_urn",
            "required": true,
            "schema": {
              "title": "Book Urn",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "page_number",
            "required": true,
            "schema": {
              "title": "Page Number",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get one page of book content.",
        "tags": ["reader"]
      }
    },
    "/api/books/{book_urn}/search": {
      "get": {
        "description": "Scan ``book_urn``'s own pages for ``q`` (the in-memory path; see\n``repositories.corpus.search_in_book``) and wrap the matches in a page.",
        "operationId": "_search_book_api_books__book_urn__search_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_urn",
            "required": true,
            "schema": {
              "title": "Book Urn",
              "type": "string"
            }
          },
          {
            "description": "Arabic query; folded before matching.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Arabic query; folded before matching.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Matches per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 30,
              "description": "Matches per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Matches to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Matches to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BookSearchMatch_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search within a book's text (diacritic- + letter-variant-insensitive).",
        "tags": ["reader"]
      }
    },
    "/api/books/{book_urn}/toc": {
      "get": {
        "description": "Return the TOC for ``book_urn`` or raise ResourceNotFoundError.",
        "operationId": "get_toc_api_books__book_urn__toc_get",
        "parameters": [
          {
            "in": "path",
            "name": "book_urn",
            "required": true,
            "schema": {
              "title": "Book Urn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Toc"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get the table of contents for a book.",
        "tags": ["reader"]
      }
    },
...
    "/api/rijal": {
      "get": {
        "description": "Wrap the repo's (slice, total) into a Page[RijalEntry] envelope.",
        "operationId": "_list_rijal_api_rijal_get",
        "parameters": [
          {
            "description": "Records per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "description": "Records per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Records to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Records to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "description": "Substring match on name / kunya / nisba.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "description": "Substring match on name / kunya / nisba.",
              "title": "Q",
              "type": "string"
            }
          },
          {
            "description": "Filter by tradition (sunni / shia / both).",
            "in": "query",
            "name": "tradition",
            "required": false,
            "schema": {
              "default": "",
              "description": "Filter by tradition (sunni / shia / both).",
              "title": "Tradition",
              "type": "string"
            }
          },
          {
            "description": "Filter by data-quality category.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "default": "",
              "description": "Filter by data-quality category.",
              "title": "Category",
              "type": "string"
            }
          },
          {
            "description": "Only entries with at least one teacher.",
            "in": "query",
            "name": "has_teachers",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only entries with at least one teacher.",
              "title": "Has Teachers",
              "type": "boolean"
            }
          },
          {
            "description": "Only entries carrying a reliability term.",
            "in": "query",
            "name": "has_reliability",
            "required": false,
            "schema": {
              "default": false,
              "description": "Only entries carrying a reliability term.",
              "title": "Has Reliability",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RijalEntry_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List rijal narrators with pagination + filters.",
        "tags": ["rijal"]
      }
    },
    "/api/rijal/{entry_id}": {
      "get": {
        "description": "Return one rijal entry by id, or raise ``ResourceNotFoundError``.",
        "operationId": "get_rijal_api_rijal__entry_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "title": "Entry Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RijalEntry"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single rijal entry by id.",
        "tags": ["rijal"]
      }
    },
...
  }
}